backend/requirements.txt
Django==4.0.0
psycopg2==2.9.9
python-dotenv==1.0.0
djangorestframework==3.14.0
tzdata==2023.3
cffi==1.16.0